All URLs now point to mitmproxy URLs
Previously, Go clients created in Deployment would point directly to the underlying
HS. This is usually fine because the Go clients are used to set up / configure the
test rather than run the test, but we have numerous helpers which convert Go clients
into actual clients-under-test. If the `.BaseURL` was not set when this happened, it
would cause the client-under-test to not be pointing to a mitmproxy'd URL.

This commit removes this footgun.