Move test setup code inside ClientSyncing blocks; support rust clients on HS2
There's two synchronisation points that need to be made to get an E2EE room
set up:
 - When Alice queries Bob's devices, Bob must have uploaded keys by then.
 - When Alice sends a message, her client must have seen Bob's join event by then.

There is a problem with ClientSyncing blocks in that it always establishes in
the order Alice, Bob, Charlie. These blocks will create and start sync loops
for those clients. When Alice is set up, there is a race condition where Bob
and Charlie may not have had their clients set up yet, causing Alice's client
to fail to see device keys for them.

This is fixed by moving all code that links Alice, Bob and Charlie together
into the ClientSyncing block, as we know for sure that all device keys are
uploaded before we even begin to create rooms. This however introduces another
race where Alice may send a message before Bob/Charlie have joined from Alice's
point of view. This is easily fixed with a Waiter that waits for Bob/Charlie's
join event to appear in the timeline.