Remove compiled rust bindings
Because we now track a moving target, it makes no sense to have this
be on some random fixed commit. Ultimately, people MUST rebuild the
rust SDK to run this project, if they want to run the rust tests. They
should be able to run the JS tests without this though.