bugfix: fix flakey TestFallbackKeyIsUsedIfOneTimeKeysRunOut
This was caused by having a too broad block on /keys/upload.
It could block Bob from uploading his OTKs, causing functions
on Bob to fail (send/recv msgs). The fix is to be more specific
with the block and only block Alice from uploading OTKs, which
is what the test intended to do.