refactor: rejig how the test suite gets updates from the browser
Previously we had a hacky system which was not formalised anywhere
in Go, relying on `JSON.stringify` and manual string concatenation.

Now we have dedicated control message types, and JS-land sends updates
as JSON objects of said control message type.

This is a pre-requisite for https://github.com/matrix-org/complement-crypto/issues/72
as that needs control messages to communicate verification state changes.