rust: always use persistence in clients
Previously we would only use persistence when the test strictly needed
it e.g to test what happens between restarts. It's preferable to always
run with persistence because:
 - it more accurately models real Element X clients e.g performance
   characteristics, runtime code.
 - any bugs in the DB layer are then also bugs in the client. Using
   the memory store has proven [error prone](https://github.com/matrix-org/matrix-rust-sdk/pull/3668)
   and fixing these bugs don't improve the stability of EX at all.

`ClientCreationOpts` will still have the `Persistence` flag though,
as it remains useful to know when a test _needs_ persistence vs when
it does not. In the future, we may clean up locally stored files
during test runtime, and this flag would allow us to know whether it
is safe to delete the files or not.