Update js-sdk import
Without this, we won't be able to run complement-crypto against an unbuilt
js-sdk (ie, a js-sdk which has been checked out from git but not yet `yarn build`-ed).

Arguably, we should import from the proper entry points, since that gives a
more representative test of how matrix-js-sdk is *meant* to be used. But that
involves adding a `yarn build` to the CI somewhere and I don't have the energy
for that right now.