Increase timeouts for SSS.
If the timeout between HS2 coming online again and Charlie on HS1 sending the message
is too short, HS1 will reject Charlie's attempt to `/keys/claim` with a HTTP 502
and "Not ready for retry". It does this because nothing cleared the 10 minute
backoff which was created when HS2 went offline. In SS, the backoff gets cleared
by Bob on HS2 hitting `/keys/query` on HS1. It does this due to receiving data
about Charlie (his join) promptly down `/sync`. This does not seem to happen in
SSS, instead the data does appear but only after the 30s timeout is reached. Therefore,
we need to wait at least 30s + some grace time for HS1 to realise HS2 is back online.