Pass in nil for the passphrase in the EnableRecovery method
The SDK always supported passphrases in addition to randomly generated
keys for the recovery subsystem, the bindings have now started to expose
this functionality as well.

Since the code-paths are the same for the passphrase based recovery
mechanism we just pass in nil and continue to test the recovery
subsystem using the randomly generated key.