Fix #150: latch closed to avoid logging after close()
We previously trusted the rust layer to not invoke callbacks
after calling `Close()`. In practice, this gets violated when
there is a sliding sync response at the same time as `Close()`
is called. Fix this by remembering when we called `Close()`
and refusing to log to stdout afterwards. Note we still log
to file.